requirements.txt
# Strands Agents SDK
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# AWS CDK for infrastructure
aws-cdk-lib>=2.170.0
constructs>=10.0.0
cdk-nag>=2.28.0

# AWS SDK
boto3>=1.42.70
botocore>=1.42.70

# Web framework (for API endpoints)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
moto[all]>=5.0.0

# Development
black>=24.10.0
ruff>=0.9.0
mypy>=1.13.0
